+++ Tests/RMPPlatformTests/TestExecutableIdentityTests.swift
@@ -12,6 +12,7 @@ struct TestExecutableIdentityTests {
         argument: "--help",
         standardOutput: """
           Usage: rmp-test put-back-race --test-run-id <uuid>
+                 rmp-test put-back-race-manual --test-run-id <uuid>
                  rmp-test [--test-run-id <uuid>] [--] <PATH>...
 
           """
@@ -24,6 +25,7 @@ struct TestExecutableIdentityTests {
         argument: "--help -a",
         standardOutput: """
           Usage: rmp-test put-back-race --test-run-id <uuid>
+                 rmp-test put-back-race-manual --test-run-id <uuid>
                  rmp-test [--test-run-id <uuid>] [--] <PATH>...
 
           """